// client/src/common/types/poi.d.ts
export interface Poi {
    id: string;
    client_id: string;
    name: string;
    use_as_address: string;
    y: string;
    x: string;
    corners: string;
    radius: string;
    geom: string;
    area: string;
    type: string;
    checkpoint_id: string;
    created_by: string;
    custom_map_id: string;
    route_id: string;
    contact: string;
    url: string;
    is_round: string;
    creation_date: number;
    changed_date: string;
    deactivated: string;
    group_id: string;
    color?: string;
}